feat(briefing): extract schools pipeline (commit 3/11) — pilot the per-pipeline contract
Workstream 6 Step 1, Commit 3/11 — pilots the per-pipeline contract on
the smallest, simplest pipeline (schools: 1 LLM call, ~165 lines extracted)
before applying it to weather/airport/news/traffic/events.

Created server/lib/briefing/pipelines/schools.js with three exports:

1. getSchoolSearchTerms(country) — synchronous helper, region-specific
   search terms (US/UK/Canada/Australia variants).

2. fetchSchoolClosures({ snapshot }) — raw LLM fetch + 15-mile distance
   filter + camelCase→snake_case normalization. Pure fetch — no SSE write,
   no cache decision. Re-exported from briefing-service.js to preserve the
   public API surface.

3. discoverSchools({ snapshot, snapshotId, cachedClosures }) — NEW pipeline
   contract. Resolves closures from cache (when orchestrator provides one)
   or fresh fetch, writes the school_closures section + fires
   CHANNELS.SCHOOL_CLOSURES pg_notify, returns { closures, reason } for
   the orchestrator's final atomic reconciliation write.

Architectural boundary: cache LOOKUP (querying other snapshots in the
same city) stays in the orchestrator — it's cross-snapshot knowledge.
Cache USAGE + fresh-fetch fallback + SSE write live in the pipeline.

In briefing-service.js:
- Removed local getSchoolSearchTerms (~20 lines) and fetchSchoolClosures
  (~125 lines)
- Replaced the inline cache-decision + fetch + writeSectionAndNotify
  block (~26 lines) with a single discoverSchools destructuring call (~8 lines)
- Re-exports fetchSchoolClosures from the new module to preserve the public
  API surface

Behavior parity: identical. Same channel, same payload shape, same write
timing relative to the 5-pipeline allSettled fan-out. Verified via
node --check on both files + dynamic import smoke test confirming the
new module exports load cleanly.

briefing-service.js: 3,329 → 3,178 lines (-151 lines net).

Per docs/review-queue/PLAN_workstream6_briefing_split.md §3.7 commit 3.
